+++ client-side/src/cpp/Backend.cpp
@@ -57,7 +57,7 @@ QString Backend::textResource(const QString &resourceName) const
 
 QString Backend::version() const
 {
-    return "1.0.7";
+    return "1.0.8";
 }
 
 void Backend::setMessage(const QString &message)